+++ src/main/java/org/folio/inventory/resources/Instances.java
@@ -1,5 +1,42 @@
 package org.folio.inventory.resources;
 
+import java.io.UnsupportedEncodingException;
+import java.lang.invoke.MethodHandles;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.net.URLEncoder;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
+import java.util.function.Consumer;
+import java.util.stream.Collectors;
+
+import org.apache.commons.lang3.StringUtils;
+import org.folio.inventory.common.WebContext;
+import org.folio.inventory.common.api.request.PagingParameters;
+import org.folio.inventory.common.domain.MultipleRecords;
+import org.folio.inventory.common.domain.Success;
+import org.folio.inventory.domain.instances.Classification;
+import org.folio.inventory.domain.instances.Contributor;
+import org.folio.inventory.domain.instances.Identifier;
+import org.folio.inventory.domain.instances.Instance;
+import org.folio.inventory.domain.instances.InstanceCollection;
+import org.folio.inventory.domain.instances.InstanceRelationship;
+import org.folio.inventory.domain.instances.InstanceRelationshipToChild;
+import org.folio.inventory.domain.instances.InstanceRelationshipToParent;
+import org.folio.inventory.domain.instances.Publication;
+import org.folio.inventory.storage.Storage;
+import org.folio.inventory.storage.external.CollectionResourceClient;
+import org.folio.inventory.support.JsonArrayHelper;
+import org.folio.inventory.support.http.client.OkapiHttpClient;
+import org.folio.inventory.support.http.client.Response;
+import org.folio.inventory.support.http.server.*;
+
 import io.vertx.core.json.Json;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
@@ -8,30 +45,9 @@ import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.handler.BodyHandler;
-import org.apache.commons.lang3.StringUtils;
-import org.folio.inventory.common.WebContext;
-import org.folio.inventory.common.api.request.PagingParameters;
-import org.folio.inventory.common.domain.MultipleRecords;
-import org.folio.inventory.domain.instances.Contributor;
-import org.folio.inventory.domain.instances.Identifier;
-import org.folio.inventory.domain.instances.Classification;
-import org.folio.inventory.domain.instances.Instance;
-import org.folio.inventory.domain.instances.InstanceCollection;
-import org.folio.inventory.storage.Storage;
-import org.folio.inventory.support.JsonArrayHelper;
-import org.folio.inventory.support.http.server.*;
-
-import java.io.UnsupportedEncodingException;
-import java.lang.invoke.MethodHandles;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
-import org.folio.inventory.domain.Metadata;
-import org.folio.inventory.domain.instances.Publication;
 
 public class Instances {
+
   private static final String INSTANCES_PATH = "/inventory/instances";
 
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
@@ -75,26 +91,28 @@ public class Instances {
 
     PagingParameters pagingParameters = PagingParameters.from(context);
 
-    if(pagingParameters == null) {
+    if (pagingParameters == null) {
       ClientErrorResponse.badRequest(routingContext.response(),
         "limit and offset must be numeric when supplied");
-
       return;
     }
 
-    if(search == null) {
+    if (search == null) {
       storage.getInstanceCollection(context).findAll(
         pagingParameters,
-        success -> JsonResponse.success(routingContext.response(),
-          toRepresentation(success.getResult(), context)),
-        FailureResponseConsumer.serverError(routingContext.response()));
-    }
-    else {
+        (Success<MultipleRecords<Instance>> success) -> {
+          makeInstancesResponse(success, routingContext, context);
+        },
+        FailureResponseConsumer.serverError(routingContext.response())
+      );
+    } else {
       try {
-        storage.getInstanceCollection(context).findByCql(search,
-          pagingParameters, success ->
-            JsonResponse.success(routingContext.response(),
-            toRepresentation(success.getResult(), context)),
+        storage.getInstanceCollection(context).findByCql(
+          search,
+          pagingParameters,
+          success -> {
+            makeInstancesResponse(success, routingContext, context);
+          },
           FailureResponseConsumer.serverError(routingContext.response()));
       } catch (UnsupportedEncodingException e) {
         ServerErrorResponse.internalError(routingContext.response(), e.toString());
@@ -107,7 +125,7 @@ public class Instances {
 
     JsonObject instanceRequest = routingContext.getBodyAsJson();
 
-    if(StringUtils.isBlank(instanceRequest.getString(Instance.TITLE_KEY))) {
+    if (StringUtils.isBlank(instanceRequest.getString(Instance.TITLE_KEY))) {
       ClientErrorResponse.badRequest(routingContext.response(),
         "Title must be provided for an instance");
       return;
@@ -117,15 +135,21 @@ public class Instances {
 
     storage.getInstanceCollection(context).add(newInstance,
       success -> {
-        try {
-          URL url = context.absoluteUrl(String.format("%s/%s",
-            INSTANCES_PATH, success.getResult().getId()));
-
-          RedirectResponse.created(routingContext.response(), url.toString());
-        } catch (MalformedURLException e) {
-          log.warn(
-            String.format("Failed to create self link for instance: %s", e.toString()));
-        }
+        Instance response = success.getResult();
+        response.setParentInstances(newInstance.getParentInstances());
+        response.setChildInstances(newInstance.getChildInstances());
+        updateInstanceRelationships(response, routingContext, context,
+                x -> {
+                    try {
+                      URL url = context.absoluteUrl(String.format("%s/%s",
+                        INSTANCES_PATH, success.getResult().getId()));
+                      RedirectResponse.created(routingContext.response(), url.toString());
+                    } catch (MalformedURLException e) {
+                      log.warn(
+                        String.format("Failed to create self link for instance: %s", e.toString()));
+                    }
+                }
+        );
       }, FailureResponseConsumer.serverError(routingContext.response()));
   }
 
@@ -140,12 +164,16 @@ public class Instances {
 
     instanceCollection.findById(routingContext.request().getParam("id"),
       it -> {
-        if(it.getResult() != null) {
+        if (it.getResult() != null) {
           instanceCollection.update(updatedInstance,
-            v -> SuccessResponse.noContent(routingContext.response()),
+            v -> {
+              updateInstanceRelationships(updatedInstance,
+                                          routingContext,
+                                          context,
+                                          (x) -> {SuccessResponse.noContent(routingContext.response());});
+            },
             FailureResponseConsumer.serverError(routingContext.response()));
-        }
-        else {
+        } else {
           ClientErrorResponse.notFound(routingContext.response());
         }
       }, FailureResponseConsumer.serverError(routingContext.response()));
@@ -154,7 +182,7 @@ public class Instances {
   private void deleteAll(RoutingContext routingContext) {
     WebContext context = new WebContext(routingContext);
 
-    storage.getInstanceCollection(context).empty (
+    storage.getInstanceCollection(context).empty(
       v -> SuccessResponse.noContent(routingContext.response()),
       FailureResponseConsumer.serverError(routingContext.response()));
   }
@@ -172,21 +200,188 @@ public class Instances {
     WebContext context = new WebContext(routingContext);
 
     storage.getInstanceCollection(context).findById(
-      routingContext.request().getParam("id"),
-      it -> {
-        if(it.getResult() != null) {
-          JsonResponse.success(routingContext.response(),
-            toRepresentation(it.getResult(), context));
-        }
-        else {
-          ClientErrorResponse.notFound(routingContext.response());
-        }
-      }, FailureResponseConsumer.serverError(routingContext.response()));
+            routingContext.request().getParam("id"),
+            it -> {
+              if (it.getResult() != null) {
+                makeInstanceResponse(it, routingContext, context);
+              } else {
+                ClientErrorResponse.notFound(routingContext.response());
+              }
+            }, FailureResponseConsumer.serverError(routingContext.response()));
   }
 
+  /**
+   * Fetch existing relationships for the instance from storage, compare them to the request. Delete, add, and modify relations as needed.
+   * @param instance The instance request containing relationship arrays to persist.
+   * @param routingContext
+   * @param context
+   */
+  private void updateInstanceRelationships(Instance instance,
+                                           RoutingContext routingContext,
+                                           WebContext context,
+                                           Consumer respond) {
+    CollectionResourceClient relatedInstancesClient = createInstanceRelationshipsClient(routingContext, context);
+    List<String> instanceId = Arrays.asList(instance.getId());
+    String query = createQueryForRelatedInstances(instanceId);
+
+    if (relatedInstancesClient != null) {
+      relatedInstancesClient.getMany(query, (Response result) -> {
+        ArrayList<CompletableFuture<Response>> allFutures = new ArrayList<>();
+        if (result.getStatusCode() == 200) {
+          JsonObject json = result.getJson();
+          List<JsonObject> relationsList = JsonArrayHelper.toList(json.getJsonArray("instanceRelationships"));
+          Map<String, InstanceRelationship> existingRelationships = new HashMap();
+          relationsList.stream().map(rel -> new InstanceRelationship(rel)).forEachOrdered(relObj -> {
+            existingRelationships.put(relObj.id, relObj);
+          });
+          Map<String, InstanceRelationship> updatingRelationships = new HashMap();
+          if (instance.getParentInstances() != null)  {
+            instance.getParentInstances().forEach(parent -> {
+              String id = (parent.id == null ? UUID.randomUUID().toString() : parent.id );
+              updatingRelationships.put(id,
+                      new InstanceRelationship(
+                              id,
+                              parent.superInstanceId,
+                              instance.getId(),
+                              parent.instanceRelationshipTypeId));
+            });
+          }
+          if (instance.getChildInstances() != null ) {
+            instance.getChildInstances().forEach(child -> {
+              String id = (child.id == null ? UUID.randomUUID().toString() : child.id );
+              updatingRelationships.put(id,
+                      new InstanceRelationship(
+                              id,
+                              instance.getId(),
+                              child.subInstanceId,
+                              child.instanceRelationshipTypeId));
+            });
+          }
+          updatingRelationships.keySet().forEach(updatingKey -> {
+            InstanceRelationship relation = updatingRelationships.get(updatingKey);
+            if (existingRelationships.containsKey(updatingKey)) {
+              if (!updatingRelationships.get(updatingKey).equals(existingRelationships.get(updatingKey))) {
+                CompletableFuture<Response> newFuture = new CompletableFuture<>();
+                allFutures.add(newFuture);
+                relatedInstancesClient.put(updatingKey, relation, newFuture::complete);
+              }
+            } else {
+              CompletableFuture<Response> newFuture = new CompletableFuture<>();
+              allFutures.add(newFuture);
+              relatedInstancesClient.post(relation, newFuture::complete);
+            }
+          });
+          existingRelationships.keySet().forEach(existingKey -> {
+            if (!updatingRelationships.containsKey(existingKey)) {
+              CompletableFuture<Response> newFuture = new CompletableFuture<>();
+              allFutures.add(newFuture);
+              relatedInstancesClient.delete(existingKey, newFuture::complete);
+            }
+          });
+          CompletableFuture.allOf(allFutures.toArray(new CompletableFuture<?>[] { }))
+                  .thenAccept(respond);
+        }
+      });
+    }
+  }
+
+
+  /**
+   * Fetches instance relationships for multiple Instance records, populates, responds
+   * @param success Multi record Instances result
+   * @param routingContext
+   * @param context
+   */
+  private void makeInstancesResponse(
+          Success<MultipleRecords<Instance>> success,
+          RoutingContext routingContext,
+          WebContext context) {
+
+    List<String> instanceIds = getInstanceIdsFromInstanceResult(success);
+    String query = createQueryForRelatedInstances(instanceIds);
+
+    try {
+      query = URLEncoder.encode(query, "UTF-8");
+    } catch (UnsupportedEncodingException e) {
+      log.error(String.format("Cannot encode query %s", query));
+    }
+    CollectionResourceClient relatedInstancesClient = createInstanceRelationshipsClient(routingContext, context);
+
+    if (relatedInstancesClient != null) {
+      relatedInstancesClient.getMany(query, (Response result) -> {
+        Map<String, List<InstanceRelationshipToParent>> parentInstanceMap = new HashMap();
+        Map<String, List<InstanceRelationshipToChild>> childInstanceMap = new HashMap();
+        if (result.getStatusCode() == 200) {
+          JsonObject json = result.getJson();
+          List<JsonObject> relationsList = JsonArrayHelper.toList(json.getJsonArray("instanceRelationships"));
+          relationsList.stream().map(rel -> {
+            addToList(childInstanceMap, rel.getString("superInstanceId"), new InstanceRelationshipToChild(rel));
+            return rel;
+          }).forEachOrdered(rel -> {
+            addToList(parentInstanceMap, rel.getString("subInstanceId"), new InstanceRelationshipToParent(rel));
+          });
+        }
+        JsonResponse.success(routingContext.response(),
+                toRepresentation(success.getResult(), parentInstanceMap, childInstanceMap, context));
+      });
+    }
+  }
+
+  /**
+   * Fetches instance relationships for a single Instance result, populates, responds
+   * @param success Single record Instance result
+   * @param routingContext
+   * @param context
+   */
+  private void makeInstanceResponse(
+          Success<Instance> success,
+          RoutingContext routingContext,
+          WebContext context) {
+
+    Instance instance = success.getResult();
+    List<String> instanceIds = getInstanceIdsFromInstanceResult(success);
+    String query = createQueryForRelatedInstances(instanceIds);
+    CollectionResourceClient relatedInstancesClient = createInstanceRelationshipsClient(routingContext, context);
+
+    if (relatedInstancesClient != null) {
+      relatedInstancesClient.getMany(query, (Response result) -> {
+        List<InstanceRelationshipToParent> parentInstanceList = new ArrayList();
+        List<InstanceRelationshipToChild> childInstanceList = new ArrayList();
+        if (result.getStatusCode() == 200) {
+          JsonObject json = result.getJson();
+          List<JsonObject> relationsList = JsonArrayHelper.toList(json.getJsonArray("instanceRelationships"));
+          relationsList.forEach(rel -> {
+            if (rel.getString(InstanceRelationship.SUPER_INSTANCE_ID_KEY).equals(instance.getId())) {
+              childInstanceList.add(new InstanceRelationshipToChild(rel));
+            } else if (rel.getString(InstanceRelationship.SUB_INSTANCE_ID_KEY).equals(instance.getId())) {
+              parentInstanceList.add(new InstanceRelationshipToParent(rel));
+            }
+          });
+        }
+        JsonResponse.success(
+                routingContext.response(),
+                toRepresentation(
+                        success.getResult(),
+                        parentInstanceList,
+                        childInstanceList,
+                        context));
+      });
+    }
+  }
+
+  /**
+   * Populates multiple Instances representation (downwards)
+   * @param wrappedInstances Set of Instances to transform to representations
+   * @param parentMap Map with list of super instances per Instance
+   * @param childMap Map with list of sub instances per Instance
+   * @param context
+   * @return
+   */
   private JsonObject toRepresentation(
-    MultipleRecords<Instance> wrappedInstances,
-    WebContext context) {
+          MultipleRecords<Instance> wrappedInstances,
+          Map<String, List<InstanceRelationshipToParent>> parentMap,
+          Map<String, List<InstanceRelationshipToChild>> childMap,
+          WebContext context) {
 
     JsonObject representation = new JsonObject();
 
@@ -194,8 +389,11 @@ public class Instances {
 
     List<Instance> instances = wrappedInstances.records;
 
-    instances.stream().forEach(instance ->
-      results.add(toRepresentation(instance, context)));
+    instances.stream().forEach(instance -> {
+      List<InstanceRelationshipToParent> parentInstances = parentMap.get(instance.getId());
+      List<InstanceRelationshipToChild> childInstances = childMap.get(instance.getId());
+      results.add(toRepresentation(instance, parentInstances, childInstances, context));
+    });
 
     representation
       .put("instances", results)
@@ -204,7 +402,18 @@ public class Instances {
     return representation;
   }
 
-  private JsonObject toRepresentation(Instance instance, WebContext context) {
+  /**
+   * Populates an Instance record representation (downwards)
+   * @param instance Instance result to transform to representation
+   * @param parentInstances Super instances for this Instance
+   * @param childInstances Sub instances for this Instance
+   * @param context
+   * @return
+   */
+  private JsonObject toRepresentation(Instance instance,
+          List<InstanceRelationshipToParent> parentInstances,
+          List<InstanceRelationshipToChild> childInstances,
+          WebContext context) {
     JsonObject resp = new JsonObject();
 
     try {
@@ -218,6 +427,8 @@ public class Instances {
     resp.put("id", instance.getId());
     resp.put(Instance.SOURCE_KEY, instance.getSource());
     resp.put(Instance.TITLE_KEY, instance.getTitle());
+    putIfNotNull(resp, Instance.PARENT_INSTANCES_KEY, parentInstances);
+    putIfNotNull(resp, Instance.CHILD_INSTANCES_KEY, childInstances);
     putIfNotNull(resp, Instance.ALTERNATIVE_TITLES_KEY, instance.getAlternativeTitles());
     putIfNotNull(resp, Instance.EDITION_KEY, instance.getEdition());
     putIfNotNull(resp, Instance.SERIES_KEY, instance.getSeries());
@@ -249,6 +460,19 @@ public class Instances {
   }
 
   private Instance requestToInstance(JsonObject instanceRequest) {
+
+    List<InstanceRelationshipToParent> parentInstances = instanceRequest.containsKey(Instance.PARENT_INSTANCES_KEY)
+      ? JsonArrayHelper.toList(instanceRequest.getJsonArray(Instance.PARENT_INSTANCES_KEY)).stream()
+          .map(json -> new InstanceRelationshipToParent(json))
+          .collect(Collectors.toList())
+      : new ArrayList<>();
+
+    List<InstanceRelationshipToChild> childInstances = instanceRequest.containsKey(Instance.CHILD_INSTANCES_KEY)
+      ? JsonArrayHelper.toList(instanceRequest.getJsonArray(Instance.CHILD_INSTANCES_KEY)).stream()
+          .map(json -> new InstanceRelationshipToChild(json))
+          .collect(Collectors.toList())
+      : new ArrayList<>();
+
     List<Identifier> identifiers = instanceRequest.containsKey(Instance.IDENTIFIERS_KEY)
       ? JsonArrayHelper.toList(instanceRequest.getJsonArray(Instance.IDENTIFIERS_KEY)).stream()
           .map(json -> new Identifier(json))
@@ -278,6 +502,8 @@ public class Instances {
       instanceRequest.getString(Instance.SOURCE_KEY),
       instanceRequest.getString(Instance.TITLE_KEY),
       instanceRequest.getString(Instance.INSTANCE_TYPE_ID_KEY))
+      .setParentInstances(parentInstances)
+      .setChildInstances(childInstances)
       .setAlternativeTitles(toListOfStrings(instanceRequest, Instance.ALTERNATIVE_TITLES_KEY))
       .setEdition(instanceRequest.getString(Instance.EDITION_KEY))
       .setSeries(toListOfStrings(instanceRequest, Instance.SERIES_KEY))
@@ -294,15 +520,55 @@ public class Instances {
       .setSourceRecordFormat(instanceRequest.getString(Instance.SOURCE_RECORD_FORMAT_KEY));
   }
 
-  private void putIfNotNull (JsonObject target, String propertyName, String value) {
-    if (value != null) target.put(propertyName, value);
+  // Utilities
+
+  private List<String> getInstanceIdsFromInstanceResult (Success success) {
+    List<String> instanceIds = new ArrayList();
+    if (success.getResult() instanceof Instance) {
+      instanceIds = Arrays.asList(((Instance)success.getResult()).getId());
+    } else if (success.getResult() instanceof MultipleRecords){
+      instanceIds = (((MultipleRecords<Instance>)success.getResult()).records.stream()
+            .map(instance -> instance.getId())
+            .filter(Objects::nonNull)
+            .distinct()
+            .collect(Collectors.toList()));
+    }
+    return instanceIds;
   }
 
-  private void putIfNotNull (JsonObject target, String propertyName, List<String> value) {
-    if (value != null) target.put(propertyName, value);
+  private CollectionResourceClient createInstanceRelationshipsClient (RoutingContext routingContext, WebContext context) {
+    CollectionResourceClient relatedInstancesClient = null;
+    try {
+      OkapiHttpClient okapiClient = createHttpClient(routingContext, context);
+      relatedInstancesClient
+      = new CollectionResourceClient(
+              okapiClient,
+              new URL(context.getOkapiLocation() + "/instance-storage/instance-relationships"));
+    } catch (MalformedURLException mfue) {
+      log.error(mfue);
+    }
+    return relatedInstancesClient;
   }
 
-  private void putIfNotNull (JsonObject target, String propertyName, Object value) {
+  private String createQueryForRelatedInstances(List<String> instanceIds) {
+    String idList = instanceIds.stream().map(String::toString).distinct().collect(Collectors.joining(" or "));
+    String query = String.format("query=(subInstanceId==(%s)+or+superInstanceId==(%s))", idList, idList);
+    return query;
+  }
+
+  private void putIfNotNull(JsonObject target, String propertyName, String value) {
+    if (value != null) {
+      target.put(propertyName, value);
+    }
+  }
+
+  private void putIfNotNull(JsonObject target, String propertyName, List<String> value) {
+    if (value != null) {
+      target.put(propertyName, value);
+    }
+  }
+
+  private void putIfNotNull(JsonObject target, String propertyName, Object value) {
     if (value != null) {
       if (value instanceof List) {
         target.put(propertyName, value);
@@ -317,4 +583,50 @@ public class Instances {
       ? JsonArrayHelper.toListOfStrings(source.getJsonArray(propertyName))
       : new ArrayList<>();
   }
+    private synchronized void addToList(Map<String, List<InstanceRelationshipToChild>> items, String mapKey, InstanceRelationshipToChild myItem) {
+    List<InstanceRelationshipToChild> itemsList = items.get(mapKey);
+
+    // if list does not exist create it
+    if (itemsList == null) {
+      itemsList = new ArrayList();
+      itemsList.add(myItem);
+      items.put(mapKey, itemsList);
+    } else {
+      // add if item is not already in list
+      if (!itemsList.contains(myItem)) {
+        itemsList.add(myItem);
+      }
+    }
+  }
+
+  private synchronized void addToList(Map<String, List<InstanceRelationshipToParent>> items, String mapKey, InstanceRelationshipToParent myItem) {
+    List<InstanceRelationshipToParent> itemsList = items.get(mapKey);
+
+    // if list does not exist create it
+    if (itemsList == null) {
+      itemsList = new ArrayList();
+      itemsList.add(myItem);
+      items.put(mapKey, itemsList);
+    } else {
+      // add if item is not already in list
+      if (!itemsList.contains(myItem)) {
+        itemsList.add(myItem);
+      }
+    }
+  }
+
+  private OkapiHttpClient createHttpClient(
+          RoutingContext routingContext,
+          WebContext context)
+          throws MalformedURLException {
+
+    return new OkapiHttpClient(routingContext.vertx().createHttpClient(),
+            new URL(context.getOkapiLocation()), context.getTenantId(),
+            context.getToken(),
+            exception -> {
+              ServerErrorResponse.internalError(routingContext.response(), String.format("Failed to contact storage module: %s",
+                      exception.toString()));
+            });
+  }
+
 }